requirements.txt
django>=4.2
requests>=2.31
django-cors-headers>=4.3
sqlalchemy>=2.0
psycopg2-binary>=2.9
bcrypt>=4.0
pyjwt>=2.8
